+++ library/src/main/java/com/chuckerteam/chucker/internal/support/ClearDatabaseService.kt
@@ -6,37 +6,19 @@ import com.chuckerteam.chucker.internal.data.repository.RepositoryProvider
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
-import java.io.Serializable
 
 internal class ClearDatabaseService : IntentService(CLEAN_DATABASE_SERVICE_NAME) {
 
     override fun onHandleIntent(intent: Intent?) {
-        when (intent?.getSerializableExtra(EXTRA_ITEM_TO_CLEAR)) {
-            is ClearAction.Transaction -> {
-                RepositoryProvider.initialize(applicationContext)
-                CoroutineScope(Dispatchers.IO).launch {
-                    RepositoryProvider.transaction().deleteAllTransactions()
-                }
-                NotificationHelper.clearBuffer()
-                NotificationHelper(this).dismissTransactionsNotification()
-            }
-            is ClearAction.Error -> {
-                RepositoryProvider.initialize(applicationContext)
-                CoroutineScope(Dispatchers.IO).launch {
-                    RepositoryProvider.throwable().deleteAllThrowables()
-                }
-                NotificationHelper(this).dismissErrorsNotification()
-            }
+        RepositoryProvider.initialize(applicationContext)
+        CoroutineScope(Dispatchers.IO).launch {
+            RepositoryProvider.transaction().deleteAllTransactions()
         }
-    }
-
-    sealed class ClearAction : Serializable {
-        object Transaction : ClearAction()
-        object Error : ClearAction()
+        NotificationHelper.clearBuffer()
+        NotificationHelper(this).dismissNotifications()
     }
 
     companion object {
         const val CLEAN_DATABASE_SERVICE_NAME = "Chucker-ClearDatabaseService"
-        const val EXTRA_ITEM_TO_CLEAR = "EXTRA_ITEM_TO_CLEAR"
     }
 }